feat: port database from Postgres to SQLite (#36)
- internal/db/db.go on modernc.org/sqlite, WAL + busy_timeout pragmas,
  SetMaxOpenConns(1)
- goose dialect "sqlite3"; squash 10 migrations into a single
  00001_initial.sql; data-only backfills dropped (window-function DELETE,
  ::jsonb backfill, archived-rescue)
- sqlc engine "sqlite"; queries moved postgres/ -> sqlite/; $N -> ?;
  ::text casts removed; IS NOT DISTINCT FROM -> IS;
  em-dashes in comments -> ASCII (caused sqlc-sqlite "SELECid" cascade
  parse errors via UTF-8 offset miscount); now() -> strftime
- internal/core/activity/list.go: ILIKE -> LIKE; search term bound twice
  because positional ? can't reuse args
- DatabaseConfig collapses to {Path}; removed DefaultDBDriver and dead
  sqlite branches in load.go, router.go, system.go doc
- internal/core/dbutil: passthrough Null helpers + NullableString +
  IsUniqueViolation against *modernc/sqlite.Error
- 3 Dockerfiles: PILOT_DATABASE_PATH=/config/pilot.db; drop pgx ENV
- Tests: pgx-based mock unique-violation rebuilt via real in-memory
  SQLite trigger; load_test password-file cases removed
- int32 -> int64 / sql.NullString -> *string / TIMESTAMPTZ -> TEXT
  ripple across ~90 sites in ~14 packages

+++ internal/core/importlist/service.go
@@ -273,9 +273,9 @@ func (s *Service) CreateExclusion(ctx context.Context, tmdbID int, title string,
 	now := time.Now().UTC().Format(time.RFC3339)
 	row, err := s.q.CreateImportExclusion(ctx, db.CreateImportExclusionParams{
 		ID:        uuid.New().String(),
-		TmdbID:    int32(tmdbID),
+		TmdbID:    int64(tmdbID),
 		Title:     title,
-		Year:      int32(year),
+		Year:      int64(year),
 		CreatedAt: now,
 	})
 	if err != nil {
@@ -360,7 +360,7 @@ func (s *Service) doSync(ctx context.Context, onlyID *string) SyncResult {
 		result.Errors = append(result.Errors, fmt.Sprintf("loading existing tmdb ids: %v", err))
 		return result
 	}
-	existingSet := make(map[int32]bool, len(existingIDs))
+	existingSet := make(map[int64]bool, len(existingIDs))
 	for _, id := range existingIDs {
 		existingSet[id] = true
 	}
@@ -371,7 +371,7 @@ func (s *Service) doSync(ctx context.Context, onlyID *string) SyncResult {
 		result.Errors = append(result.Errors, fmt.Sprintf("loading exclusions: %v", err))
 		return result
 	}
-	excludedSet := make(map[int32]bool, len(excludedIDs))
+	excludedSet := make(map[int64]bool, len(excludedIDs))
 	for _, id := range excludedIDs {
 		excludedSet[id] = true
 	}
@@ -406,7 +406,7 @@ func (s *Service) doSync(ctx context.Context, onlyID *string) SyncResult {
 			if item.TMDbID == 0 {
 				continue
 			}
-			tmdbID := int32(item.TMDbID)
+			tmdbID := int64(item.TMDbID)
 			if existingSet[tmdbID] {
 				result.SeriesSkipped++
 				continue
